Remove any charm code from tests as suggested by review
