+++ tests/integration/testers/any_charm_src/ipa/any_charm.py
@@ -3,6 +3,7 @@
 
 """any-charm src-overwrite for a simple ingress-per-app requirer."""
 
+import json
 import os
 import sys
 
@@ -18,3 +19,23 @@ class AnyCharm(AnyCharmBase):
         self.ipa = IngressPerAppRequirer(
             self, host="foo.bar", port=80, relation_name="require-ingress"
         )
+
+    def get_relation_data(self):
+        rel = self.ipa.relation
+        if rel is None:
+            return {"url": None, "app_data": {}, "unit_data": {}}
+
+        def _decode(d):
+            result = {}
+            for k, v in d.items():
+                try:
+                    result[k] = json.loads(v)
+                except (json.JSONDecodeError, TypeError):
+                    result[k] = v
+            return result
+
+        return {
+            "url": self.ipa.url,
+            "app_data": _decode(dict(rel.data[self.app])),
+            "unit_data": _decode(dict(rel.data[self.unit])),
+        }
